Move EM27Property out of hardware module
