+++ finesse/hardware/plugins/em27/em27_scraper.py
@@ -2,7 +2,6 @@
 
 This is used to scrape the PSF27Sensor data table off the server.
 """
-from dataclasses import dataclass
 from decimal import Decimal
 from functools import partial
 
@@ -11,39 +10,8 @@ from PySide6.QtCore import Slot
 from PySide6.QtNetwork import QNetworkAccessManager, QNetworkReply, QNetworkRequest
 
 from finesse.config import EM27_URL
+from finesse.em27_info import EM27Property
 from finesse.hardware.pubsub_decorators import pubsub_broadcast
-
-
-@dataclass
-class EM27Property:
-    """Class for representing EM27 monitored properties.
-
-    Args:
-        name: name of the physical quantity
-        value: value of the physical quantity
-        unit: unit in which the value is presented
-    """
-
-    name: str
-    value: Decimal
-    unit: str
-
-    def __str__(self) -> str:
-        """Print a property's name, value and unit in a readable format.
-
-        Returns:
-            str: The name, value and unit of a property.
-        """
-        return f"{self.name} = {self.value:.6f} {self.unit}"
-
-    def val_str(self) -> str:
-        """Print a property's value and unit in required format.
-
-        Returns:
-            str: The value and unit of a property in the format consistent with
-                 the previous FINESSE GUI.
-        """
-        return f"{self.value:.6f} {self.unit}"
 
 
 def get_em27sensor_data(content: str) -> list[EM27Property]:
