+++ finesse/hardware/plugins/em27/opus_interface.py
@@ -14,7 +14,7 @@ from PySide6.QtCore import Slot
 from PySide6.QtNetwork import QNetworkAccessManager, QNetworkReply, QNetworkRequest
 
 from finesse.config import OPUS_IP
-from finesse.em27_status import EM27Status
+from finesse.em27_info import EM27Status
 
 from .opus_interface_base import OPUSInterfaceBase
